refactor: split tool_registry.py into domain-specific registry modules (#140)
