+++ src/mcp_server_git/lean/registry_github.py
@@ -1,0 +1,564 @@
+"""GitHub domain tool registration for lean MCP interface."""
+
+import logging
+from typing import Any
+
+from ..github import api as github_ops
+from ..github.models import (
+    GitHubBulkUpdateIssues,
+    GitHubCreateIssue,
+    GitHubCreateIssueFromTemplate,
+    GitHubCreateRelease,
+    GitHubCreateRepo,
+    GitHubDeleteBranchProtection,
+    GitHubDeleteRelease,
+    GitHubDeleteReleaseAsset,
+    GitHubDisableAutomatedSecurityFixes,
+    GitHubDisableVulnerabilityAlerts,
+    GitHubEditPRDescription,
+    GitHubEnableAutomatedSecurityFixes,
+    GitHubEnableVulnerabilityAlerts,
+    GitHubGetActionsPermissions,
+    GitHubGetAutomatedSecurityFixes,
+    GitHubGetBranchProtection,
+    GitHubGetFailingJobs,
+    GitHubGetIssue,
+    GitHubGetJobLogs,
+    GitHubGetPRChecks,
+    GitHubGetPRDetails,
+    GitHubGetPRFiles,
+    GitHubGetPRStatus,
+    GitHubGetRelease,
+    GitHubGetRepoSettings,
+    GitHubGetSecurityAnalysis,
+    GitHubGetVulnerabilityAlerts,
+    GitHubGetWorkflowPermissions,
+    GitHubGetWorkflowRun,
+    GitHubListIssues,
+    GitHubListPullRequests,
+    GitHubListReleaseAssets,
+    GitHubListReleases,
+    GitHubListWorkflowRuns,
+    GitHubSearchIssues,
+    GitHubUpdateActionsPermissions,
+    GitHubUpdateBranchProtection,
+    GitHubUpdateIssue,
+    GitHubUpdateRelease,
+    GitHubUpdateRepoSettings,
+    GitHubUpdateWorkflowPermissions,
+    GitHubUploadReleaseAsset,
+)
+from .interface import ToolDefinition
+
+logger = logging.getLogger(__name__)
+
+
+def _register_github_tools(interface: Any, github_service: Any):
+    """Register all GitHub domain tools."""
+
+    github_tools = [
+        # PR Tools
+        ToolDefinition(
+            name="github_get_pr_checks",
+            implementation=github_ops.github_get_pr_checks,
+            description="Get check runs for a pull request",
+            schema=GitHubGetPRChecks.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_failing_jobs",
+            implementation=github_ops.github_get_failing_jobs,
+            description="Get detailed information about failing CI jobs for a pull request",
+            schema=GitHubGetFailingJobs.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_pr_details",
+            implementation=github_ops.github_get_pr_details,
+            description="Get detailed information about a pull request",
+            schema=GitHubGetPRDetails.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_list_pull_requests",
+            implementation=github_ops.github_list_pull_requests,
+            description="List pull requests with filtering options",
+            schema=GitHubListPullRequests.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_pr_status",
+            implementation=github_ops.github_get_pr_status,
+            description="Get the status of a pull request",
+            schema=GitHubGetPRStatus.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_pr_files",
+            implementation=github_ops.github_get_pr_files,
+            description="Get files changed in a pull request",
+            schema=GitHubGetPRFiles.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_edit_pr_description",
+            implementation=github_ops.github_edit_pr_description,
+            description="Edit the description of a pull request",
+            schema=GitHubEditPRDescription.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        # Workflow Tools
+        ToolDefinition(
+            name="github_get_workflow_run",
+            implementation=github_ops.github_get_workflow_run,
+            description="Get detailed workflow run information",
+            schema=GitHubGetWorkflowRun.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_list_workflow_runs",
+            implementation=github_ops.github_list_workflow_runs,
+            description="List workflow runs for a repository with comprehensive filtering",
+            schema=GitHubListWorkflowRuns.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_job_logs",
+            implementation=github_ops.github_get_job_logs,
+            description="Get logs for a specific GitHub Actions job for CI debugging",
+            schema=GitHubGetJobLogs.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        # Issue Tools
+        ToolDefinition(
+            name="github_create_issue",
+            implementation=github_ops.github_create_issue,
+            description="Create a new GitHub issue",
+            schema=GitHubCreateIssue.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_list_issues",
+            implementation=github_ops.github_list_issues,
+            description="List GitHub issues with filtering options",
+            schema=GitHubListIssues.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_issue",
+            implementation=github_ops.github_get_issue,
+            description="Get a single GitHub issue by number with full details",
+            schema=GitHubGetIssue.model_json_schema(),
+            domain="github",
+            complexity="core",
+        ),
+        ToolDefinition(
+            name="github_update_issue",
+            implementation=github_ops.github_update_issue,
+            description="Update an existing GitHub issue",
+            schema=GitHubUpdateIssue.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_search_issues",
+            implementation=github_ops.github_search_issues,
+            description="Search GitHub issues with advanced query capabilities",
+            schema=GitHubSearchIssues.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_create_issue_from_template",
+            implementation=github_ops.github_create_issue_from_template,
+            description="Create issue from template",
+            schema=GitHubCreateIssueFromTemplate.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_bulk_update_issues",
+            implementation=github_ops.github_bulk_update_issues,
+            description="Bulk update multiple issues",
+            schema=GitHubBulkUpdateIssues.model_json_schema(),
+            domain="github",
+            complexity="comprehensive",
+        ),
+        # Additional GitHub tools from CLI (7 more to reach 22 total)
+        ToolDefinition(
+            name="github_create_pr",
+            implementation=github_ops.github_create_pr,
+            description="Create a new pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "title": {"type": "string"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_merge_pr",
+            implementation=github_ops.github_merge_pr,
+            description="Merge a pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_add_pr_comment",
+            implementation=github_ops.github_add_pr_comment,
+            description="Add a comment to a pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                    "body": {"type": "string"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_pr_comments",
+            implementation=github_ops.github_get_pr_comments,
+            description="Get top-level conversation comments on a pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_pr_reviews",
+            implementation=github_ops.github_get_pr_reviews,
+            description="Get inline code review comments on a pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_reply_to_pr_comment",
+            implementation=github_ops.github_reply_to_pr_comment,
+            description="Reply to a specific review comment thread on a pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                    "comment_id": {"type": "integer"},
+                    "body": {"type": "string"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_close_pr",
+            implementation=github_ops.github_close_pr,
+            description="Close a pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_reopen_pr",
+            implementation=github_ops.github_reopen_pr,
+            description="Reopen a closed pull request",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_update_pr",
+            implementation=github_ops.github_update_pr,
+            description="Update a pull request (title, body, state, or base)",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "pr_number": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_await_workflow_completion",
+            implementation=github_ops.github_await_workflow_completion,
+            description="Monitor a GitHub Actions workflow run until completion",
+            schema={
+                "type": "object",
+                "properties": {
+                    "repo_owner": {"type": "string"},
+                    "repo_name": {"type": "string"},
+                    "run_id": {"type": "integer"},
+                },
+            },
+            domain="github",
+            complexity="focused",
+        ),
+        # Repository Settings Management (Issue #41)
+        ToolDefinition(
+            name="github_get_repo_settings",
+            implementation=github_ops.github_get_repo_settings,
+            description="Get repository settings including features, merge options, and security settings",
+            schema=GitHubGetRepoSettings.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_update_repo_settings",
+            implementation=github_ops.github_update_repo_settings,
+            description="Update repository settings (merge strategies, features, security)",
+            schema=GitHubUpdateRepoSettings.model_json_schema(),
+            domain="github",
+            complexity="comprehensive",
+        ),
+        # GitHub Actions Configuration (Issue #41)
+        ToolDefinition(
+            name="github_get_actions_permissions",
+            implementation=github_ops.github_get_actions_permissions,
+            description="Get GitHub Actions permissions for a repository",
+            schema=GitHubGetActionsPermissions.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_update_actions_permissions",
+            implementation=github_ops.github_update_actions_permissions,
+            description="Update GitHub Actions permissions (enable/disable, allowed actions)",
+            schema=GitHubUpdateActionsPermissions.model_json_schema(),
+            domain="github",
+            complexity="comprehensive",
+        ),
+        ToolDefinition(
+            name="github_get_workflow_permissions",
+            implementation=github_ops.github_get_workflow_permissions,
+            description="Get default workflow permissions (GITHUB_TOKEN permissions)",
+            schema=GitHubGetWorkflowPermissions.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_update_workflow_permissions",
+            implementation=github_ops.github_update_workflow_permissions,
+            description="Update default workflow permissions and PR approval settings",
+            schema=GitHubUpdateWorkflowPermissions.model_json_schema(),
+            domain="github",
+            complexity="comprehensive",
+        ),
+        # Branch Protection Rules (Issue #41)
+        ToolDefinition(
+            name="github_get_branch_protection",
+            implementation=github_ops.github_get_branch_protection,
+            description="Get branch protection rules for a specific branch",
+            schema=GitHubGetBranchProtection.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_update_branch_protection",
+            implementation=github_ops.github_update_branch_protection,
+            description="Create or update branch protection rules (status checks, reviews, restrictions)",
+            schema=GitHubUpdateBranchProtection.model_json_schema(),
+            domain="github",
+            complexity="comprehensive",
+        ),
+        ToolDefinition(
+            name="github_delete_branch_protection",
+            implementation=github_ops.github_delete_branch_protection,
+            description="Delete branch protection rules",
+            schema=GitHubDeleteBranchProtection.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        # Security & Compliance (Issue #41)
+        ToolDefinition(
+            name="github_get_vulnerability_alerts",
+            implementation=github_ops.github_get_vulnerability_alerts,
+            description="Check if vulnerability alerts (Dependabot alerts) are enabled",
+            schema=GitHubGetVulnerabilityAlerts.model_json_schema(),
+            domain="github",
+            complexity="core",
+        ),
+        ToolDefinition(
+            name="github_enable_vulnerability_alerts",
+            implementation=github_ops.github_enable_vulnerability_alerts,
+            description="Enable vulnerability alerts (Dependabot alerts) for a repository",
+            schema=GitHubEnableVulnerabilityAlerts.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_disable_vulnerability_alerts",
+            implementation=github_ops.github_disable_vulnerability_alerts,
+            description="Disable vulnerability alerts (Dependabot alerts) for a repository",
+            schema=GitHubDisableVulnerabilityAlerts.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_automated_security_fixes",
+            implementation=github_ops.github_get_automated_security_fixes,
+            description="Check if automated security fixes (Dependabot security updates) are enabled",
+            schema=GitHubGetAutomatedSecurityFixes.model_json_schema(),
+            domain="github",
+            complexity="core",
+        ),
+        ToolDefinition(
+            name="github_enable_automated_security_fixes",
+            implementation=github_ops.github_enable_automated_security_fixes,
+            description="Enable automated security fixes (Dependabot security updates)",
+            schema=GitHubEnableAutomatedSecurityFixes.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_disable_automated_security_fixes",
+            implementation=github_ops.github_disable_automated_security_fixes,
+            description="Disable automated security fixes (Dependabot security updates)",
+            schema=GitHubDisableAutomatedSecurityFixes.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_security_analysis",
+            implementation=github_ops.github_get_security_analysis,
+            description="Get comprehensive security analysis status (vulnerability alerts, security fixes, secret scanning)",
+            schema=GitHubGetSecurityAnalysis.model_json_schema(),
+            domain="github",
+            complexity="comprehensive",
+        ),
+        # Release Management Tools
+        ToolDefinition(
+            name="github_create_release",
+            implementation=github_ops.github_create_release,
+            description="Create a new GitHub release with tag, title, and notes",
+            schema=GitHubCreateRelease.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_get_release",
+            implementation=github_ops.github_get_release,
+            description="Get release information by ID or tag name",
+            schema=GitHubGetRelease.model_json_schema(),
+            domain="github",
+            complexity="core",
+        ),
+        ToolDefinition(
+            name="github_list_releases",
+            implementation=github_ops.github_list_releases,
+            description="List releases for a repository with pagination",
+            schema=GitHubListReleases.model_json_schema(),
+            domain="github",
+            complexity="core",
+        ),
+        ToolDefinition(
+            name="github_update_release",
+            implementation=github_ops.github_update_release,
+            description="Update an existing release's properties",
+            schema=GitHubUpdateRelease.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_delete_release",
+            implementation=github_ops.github_delete_release,
+            description="Delete a GitHub release (does not delete the git tag)",
+            schema=GitHubDeleteRelease.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        ToolDefinition(
+            name="github_upload_release_asset",
+            implementation=github_ops.github_upload_release_asset,
+            description="Upload a file as an asset to a GitHub release",
+            schema=GitHubUploadReleaseAsset.model_json_schema(),
+            domain="github",
+            complexity="advanced",
+        ),
+        ToolDefinition(
+            name="github_list_release_assets",
+            implementation=github_ops.github_list_release_assets,
+            description="List assets attached to a release",
+            schema=GitHubListReleaseAssets.model_json_schema(),
+            domain="github",
+            complexity="core",
+        ),
+        ToolDefinition(
+            name="github_delete_release_asset",
+            implementation=github_ops.github_delete_release_asset,
+            description="Delete an asset from a release",
+            schema=GitHubDeleteReleaseAsset.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+        # Repository Creation (Issue #127)
+        ToolDefinition(
+            name="github_create_repo",
+            implementation=github_ops.github_create_repo,
+            description="Create a new GitHub repository (personal or organization)",
+            schema=GitHubCreateRepo.model_json_schema(),
+            domain="github",
+            complexity="focused",
+        ),
+    ]
+
+    for tool in github_tools:
+        interface.register_tool(tool)
+
+    logger.info(f"Registered {len(github_tools)} GitHub tools")
